closes #60

+++ api/Models/WebMap/Symbol.cs
@@ -8,7 +8,7 @@ namespace daq_api.Models.WebMap
         public Outline Outline { get; set; }
         public string Type { get; set; }
         public string Style { get; set; }
-        public int? Size { get; set; }
+        public double? Size { get; set; }
         public int? Angle { get; set; }
         public int? Xoffset { get; set; }
         public int? Yoffset { get; set; }
